python-api/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
gspread==6.1.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
gunicorn==22.0.0
requests==2.31.0
facebook-business==19.0.0
google-ads==21.3.0
psycopg2-binary==2.9.9
